refactor : change automatic login checkpoint
로그인 화면에서 앱 로딩으로 자동 로그인 체크 시점 변경.
